requirements.txt
streamlit>=1.35.0
requests>=2.32.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.40.0
pydantic>=2.8.0
